fix(popup): demo拆解与规范
