+++ src/packages/popup/demos/taro/demo1.tsx
@@ -1,5 +1,6 @@
 import React, { useState } from 'react'
 import { Popup, Cell } from '@nutui/nutui-react-taro'
+import { View } from '@tarojs/components'
 
 const Demo1 = () => {
   const [showBasic, setShowBasic] = useState(false)
@@ -18,7 +19,15 @@ const Demo1 = () => {
           setShowBasic(false)
         }}
       >
-        正文
+        <View style={{ height: '200px', overflowY: 'scroll' }}>
+          {Array.from({ length: 20 })
+            .fill('')
+            .map((_, i) => (
+              <Cell key={i}>
+                <View>正文</View>
+              </Cell>
+            ))}
+        </View>
       </Popup>
     </>
   )
