+++ src/packages/popup/demos/taro/demo4.tsx
@@ -1,25 +1,47 @@
 import React, { useState } from 'react'
+import { Failure } from '@nutui/icons-react-taro'
 import { Popup, Cell } from '@nutui/nutui-react-taro'
 
 const Demo4 = () => {
-  const [showBottomRound, setShowBottomRound] = useState(false)
+  const [showOverlayStop, setShowOverlayStop] = useState(false)
+  const [showCloseIconStop, setShowCloseIconStop] = useState(false)
 
   return (
     <>
       <Cell
-        title="圆角弹框"
+        title="阻塞点击 Overlay 关闭"
         onClick={() => {
-          setShowBottomRound(true)
+          setShowOverlayStop(true)
+        }}
+      />
+      <Popup
+        visible={showOverlayStop}
+        style={{ padding: '30px 50px' }}
+        onClose={() => {
+          setShowOverlayStop(false)
+        }}
+        onOverlayClick={() => {
+          console.log('onOverlayClick')
+          return false
+        }}
+      >
+        正文
+      </Popup>
+      <Cell
+        title="阻塞点击 close icon 关闭"
+        onClick={() => {
+          setShowCloseIconStop(true)
         }}
       />
       <Popup
         closeable
-        visible={showBottomRound}
-        style={{ height: '20%' }}
+        closeIcon={<Failure size={12} />}
+        visible={showCloseIconStop}
+        closeOnOverlayClick={false}
+        style={{ height: '40%' }}
         position="bottom"
-        round
-        onClose={() => {
-          setShowBottomRound(false)
+        onCloseIconClick={() => {
+          console.log('onCloseIconClick')
         }}
       />
     </>
